Extract connectWallet
